0510 프로그래머스 연습문제

+++ JavaPractice/Programmers/spaceXNumbern.java
@@ -10,7 +10,7 @@ public class spaceXNumbern {
         long[] answer =new long[n];
         
         for(int i=0;i<n;i++){
-            answer[i]=x*(i+1);
+            answer[i]=(long)x*(i+1);
             System.out.print(answer[i]);
         }
